Added notify to solvers
