+++ crates/solvers/src/api/routes/mod.rs
@@ -1,0 +1,4 @@
+mod notify;
+mod solve;
+
+pub(super) use {notify::notify, solve::solve};
