+++ crates/solvers/src/api/routes/solve/dto/mod.rs
