+++ crates/solvers/src/api/routes/solve/dto/solution.rs
