+++ crates/solvers/src/api/routes/solve/mod.rs
@@ -1,0 +1,46 @@
+use tracing::Instrument;
+
+mod dto;
+
+use {crate::domain::solver::Solver, std::sync::Arc};
+
+pub async fn solve(
+    state: axum::extract::State<Arc<Solver>>,
+    auction: axum::extract::Json<dto::Auction>,
+) -> (
+    axum::http::StatusCode,
+    axum::response::Json<dto::Response<dto::Solutions>>,
+) {
+    let handle_request = async {
+        let auction = match auction.to_domain() {
+            Ok(value) => value,
+            Err(err) => {
+                tracing::warn!(?err, "invalid auction");
+                return (
+                    axum::http::StatusCode::BAD_REQUEST,
+                    axum::response::Json(dto::Response::Err(err)),
+                );
+            }
+        };
+
+        tracing::trace!(?auction);
+
+        let auction_id = auction.id;
+        let solutions = state
+            .solve(auction)
+            .instrument(tracing::info_span!("auction", id = %auction_id))
+            .await;
+
+        tracing::trace!(?solutions);
+
+        let solutions = dto::Solutions::from_domain(&solutions);
+        (
+            axum::http::StatusCode::OK,
+            axum::response::Json(dto::Response::Ok(solutions)),
+        )
+    };
+
+    handle_request
+        .instrument(tracing::info_span!("/solve"))
+        .await
+}
